naming conventions and updating auto plan naming
